RE: [Raw] "leg" vs "lane"
Hello Rahul

Many thanks for those useful comments. Let’s do a pass cleaning up the DAO projection draft and aligning to RAW.
-	subTrack is removed
-	leg -> Lane
-	define stand-alone as opposed to stand-alone Track.

To your points below:

	1. Section 2.2 in DAO projection -29 says, "The RAW Architecture also defines the concepts of subTrack, Segment, leg, East-West, and more." but I did not find any context of "Leg" or "subTrack" in that draft. Maybe we should fix this while at it.

This takes a little rewrite to explain the concepts well, and compare what RPL DODAGs are vs. Tracks. Proposed update:

“
2.2.  References

   In this document, readers will encounter terms and concepts that are
   discussed in the "Routing Protocol for Low Power and Lossy Networks"
   [RPL], the "6TiSCH Architecture" [RFC9030], the "Deterministic
   Networking Architecture" [RFC8655], the "Reliable and Available
   Wireless (RAW) Architecture" [RAW-ARCHI], and "Terminology in Low
   power And Lossy Networks" [RFC7102].  Both architecture documents
   define the concept of Track in a compatible fashion.  This documents
   only builds Tracks that are DODAGs, meaning that all links are
   oriented From Ingress to Egress.  This specification also utilizes
   the terms Segment and Lane that are also defined in the RAW
   Architecture.

   As opposed to routing trees, RPL DODAGs are typically constructed to
   provide redundancy and dynamically adapt the forwarding operation to
   the state of the LLN links.  Note that the plain forwarding operation
   over DODAGs does not provide redundancy for all nodes, since at least
   the node nearest to the Root does not have an alternate feasible
   successor.

   RAW solves that problem by defining Protection Paths that can be
   fully non-congruent and can be activated dynamically upon failures.
   This requires additional control to take the routing decision early
   enough along the Track to route around the failure.

   RAW only uses single-ended DODAGs, meaning that they can be reversed
   in another DODAG by reversing all the links.  The Ingress of the
   Track is the Root of the DODAG, whereas the Egress is the Root of the
   reversed DODAG.  From the RAW perspective, single-ended DODAGs are
   special Tracks that only have East-West links, and that can be
   leveraged to provide Protection services by defining destination-
   oriented Protection Paths within the DODAG.

“

	2. Section 2.4.5.5 defines "Stand-Alone Track" ... but the remainder of the text uses the term "Stand-Alone Segment". There is no reference of "Stand-Alone Track" except in the terms section.

My mistake; a Track is “serial” not “stand alone”. What’s “stand alone” is a Segment or a Lane that forms a Serial Track with one P DAO. For a Lane that means it is strict as opposed to loose. For a segùment it means it is end to end.
“

2.4.5.5.  Stand-Alone

   Refers to a Segment or a Lane that is installed with a single P-DAO
   and fully defines a Serial Track, e.g., a Stand-Alone Segment
   installed with a single Storing Mode Via Information option (SM-VIO)
   all the way between Ingress and Egress.
“

	3. In all other contexts, the lane can be a suitable alternative to leg.

Very cool!

Many thanks Rahul 😊

All the best,

Pascal

From: Rahul Jadhav <rahul.ietf@gmail.com>
Sent: mercredi 7 décembre 2022 4:20
To: Pascal Thubert (pthubert) <pthubert@cisco.com>
Cc: ROLL WG (roll@ietf.org) <roll@ietf.org>; Don Fedyk <dfedyk@labn.net>; raw@ietf.org
Subject: Re: [Raw] "leg" vs "lane"

The term leg appears in existing literature (last I saw it getting used was for SIP Call legs) and hence I too think it might be better to use the term "lane" in the context of DAO projection. Would like to know if there was a discussion/background on RAW/elsewhere about how the term lane originated.

After going through ver -29 of DAO projection (simply grepping for "leg" and checking if "lane" makes more sense), here are some thoughts:
1. Section 2.2 in DAO projection -29 says, "The RAW Architecture also defines the concepts of subTrack, Segment, leg, East-West, and more." but I did not find any context of "Leg" or "subTrack" in that draft. Maybe we should fix this while at it.
2. Section 2.4.5.5 defines "Stand-Alone Track" ... but the remainder of the text uses the term "Stand-Alone Segment". There is no reference of "Stand-Alone Track" except in the terms section.
3. In all other contexts, the lane can be a suitable alternative to leg.

Regards,
Rahul

On Tue, 6 Dec 2022 at 12:52, Pascal Thubert (pthubert) <pthubert=40cisco.com@dmarc.ietf.org> wrote:
Dear all

RAW has mostly completed its architecture work, for which the DAO projection work provides a useful component. We are in the process of cleaning up the terminology, juggling with the tension between compatibility with the IETF art when fit and avoiding confusion by overloading existing terms too much.

As part of the review, we got a comment that the term "lane" could be appropriate for east west protection paths.
The DAO projection draft introduced the term "leg" for this.

I agree with the commenter that "lane" is a good term (probably better than "leg") and would like to sync RAW and ROLL on it, using the RAW architecture as the defining place for the concept.

Please let me know if that's a bad idea, otherwise it will show in the next publication.

All the best,

Pascal

--
RAW mailing list
RAW@ietf.org
https://www.ietf.org/mailman/listinfo/raw