From: Roll <roll-bounces@ietf.org> On Behalf Of Alvaro Retana Sent: mercredi 4 janvier 2023 20:02 To: mariainesrobles@googlemail.com Cc: roll@ietf.org; roll-chairs@ietf.org Subject: [SUSPECTED SPAM] Re: [Roll] Publication has been requested for draft-ietf-roll-dao-projection-31
Hi!

I had a very quick look at this document.  It needs additional reviews.

(1) §4 uses "amends" as an indication of a formal Update (along the lines of draft-kuehlewind-rswg-updates-tag).  Because what we have today is only the Updates tag, please indicate in the header and the text any Updated RFCs.

(1a) One of them seems to be rfc6553, so we'll need to get 6man to review that.

(2) §3.7.2 (Positioning vs. Related IETF Standards) mentions specific expectations related to the work of other WGs:

    3.7.2.1.  Extending 6TiSCH
    3.7.2.2.  Mapping to DetNet
    3.7.2.3.  Leveraging PCE
    3.7.2.4.  Providing for RAW

I couldn't find a (formal) request for these WGs to review the document.  Did I miss it?  If not, then please request it.  I know that there is significant participation overlap -- that is not enough.

(3) Per the request in October [1], I need you to also request a rtg-dir review.

While I'm hoping that none of these reviews result in significant changes, I'm returning the document to the WG.

Suggestion:  Once the draft has been revised with the Updates information, issue a new WGLC and cc all the relevant WGs on it.

Thanks!

Alvaro.

[1] Sent to rtg-chairs (which is not a proper mailing-list), search for "Message-ID: <AM7PR03MB6451605AB56AFC67424797A3EE229@AM7PR03MB6451.eurprd03.prod.outlook.com>".

On January 4, 2023 at 11:48:16 AM, Ines Robles via Datatracker (noreply@ietf.org) wrote: