Hello Ran:
many thanks for extracting those cases; let's see:

Le jeu. 6 mars 2025 à 04:40, <chen.ran@zte.com.cn> a écrit :
Hi John,

Thank you very much for pointing this out. I completely agree that they should be applied thoughtfully rather than indiscriminately, However, I think capitalization is more appropriate in the following cases.

-But the RPL routing information headers may not be supported on all type of routed network infrastructures, especially not in high-speed routers.

not for that one, I actually meant lower case, just stating the art of the Internet.

-The RPL Root must add a source routing header to all downward packets.

This is specified by RFC 6550 as opposed to this spec, I use lowercase in that situation.

-If the 'K' Flag is present in the P-DAO, the P-DAO must be acknowledged using a DAO-ACK that is sent back to the address of the Root from which the P-DAO was received.

yep, thanks for the catch!

-A P-DAO that creates or updates a protection path MUST be sent to a GUA or a ULA of the Ingress of the protection path; it must contain the full list of hops in the protection path unless the protection path is being removed.

yes again

-The router MAY install additional routes towards the Via Addresses that appear in the SM-VIO after its own address, if any, but in case of a conflict or a lack of resource, the route(s) to the Target(s) are the ones that must be installed in priority.

OK

-When injecting a packet in a Track, the Ingress router must encapsulate the packet using IP-in-IP to add the Source Routing Header with the final destination set to the Track Egress.

Then again this is RFC 6550

-With this specification, only the Root may generate P-DAO messages. PDR messages may only be sent to the Root.

OK, with the caveat that we do not have MAY ONLY. Let me reword as:
"
   With this specification, the Root MAY generate P-DAO messages but
   other nodes MUST NOT do so. PDR messages MUST be sent to the Root.
 "

-That trust should propagate from Egress to Ingress in the case of a Storing Mode P-DAO.

I do not have a spec for this to happen so it's only a general recommendation.

Works?

all the best,

Pascal